feat: setup sdk
This PR introduces a base setup of the SDK which we'll use to interact with the Solana Program and our server.

For now, it only exposes a single method, estimateFees to compute the fee to be paid in SOL based on teh file size and the duration it should be stored for.

For now, we need to properly define or provide(?) a IDL and also the Program Derived Address before i can move on to the createDeposit method.

The build is also currently set to cater for two environments, the browser and node. Need to know if we should make a couple of the dependencies external or since the consumers of the sdk would most-likely have the packages/dependencies installed in their projects.

This would help us ensure we have a lean build.