feat: added backend solana deposit route.
set up the deposit route for program txns.

for now i'm reading the program keypair read locally generated from the anchor build. i also encountered a lot of issues while setting up the program for teh first time, and i've documented them in the README so it is easier to walkthrough in the future.

the deposit route is pretty much a replica of what the sdk should do. but since we're heading in a new direction now, where the backend handles communication with the program and the sdk does I/O with the backend, we'll update most of the code there.

i'll need to speak with devansh on the essence of the required ADMIN_API_KEY too.