fix: defer blockhash and transaction construction to the sdk/client
when i tried testing by making a test deposit with a sample wallet, the blockhas was always invalidated before the client tries to use it for the transaction. i'll keep digging into this.

i've also stripped the usage of borsh-xyz as it led to a couple of complicated issues serializing the transaction layout