feat: added backend solana deposit route. (#10)
* feat: added backend solana deposit route.

set up the deposit route for program txns.

for now i'm reading the program keypair read locally generated from the anchor build. i also encountered a lot of issues while setting up the program for teh first time, and i've documented them in the README so it is easier to walkthrough in the future.

the deposit route is pretty much a replica of what the sdk should do. but since we're heading in a new direction now, where the backend handles communication with the program and the sdk does I/O with the backend, we'll update most of the code there.

i'll need to speak with devansh on the essence of the required ADMIN_API_KEY too.

* fix: defer blockhash and transaction construction to the sdk/client

when i tried testing by making a test deposit with a sample wallet, the blockhas was always invalidated before the client tries to use it for the transaction. i'll keep digging into this.

i've also stripped the usage of borsh-xyz as it led to a couple of complicated issues serializing the transaction layout