fix: compute file CID with ipfs-car before upload
this prevents the edge case where people can upload a file without paying a deposit. now we compute the CID of the file before uploading it — quite similar to how CIDs are created in Storacha from the uploadFile/UploadDirectory methods

the only potential blocker now is when people try to upload a folder instead of a file. i'm not sure if the CID would match (or the process would work at all, since it uses uploadFile for now. we'll cover that later though. need to get the PoC working first).

this pull request also led to a slight change with managing delegations as opposed to teh fomer approach where the client/sdk expects people to supply a StorachaKey and proof variables.