fix: storacha client should not be init on the backend
the former approach where we used the storacha client on the server centralizes data storage our own pre-owned space which we've established by providing a hardcoded proof for delegations, and a storahca key (supplied via environment variables).

this takes away the control from users owning their spaces and having control over what they upload.

in the future, we may decide to ask them to delegate access to us, if they choose to. but for now, the okay approach is to go with the "user-owned" architecture option where people register a space and upload directly to the service.