feat: expose client sdk via a react hook (#29)
* feat: expose client sdk via a react hook

this is a refactor to improve the existing sdk implementation bringing it to a point where it is usable by devs.

some of the changes include updates to the types which we'll emit in a declaration file via our bundler (tsup), removal of redundant files, and ensuring that an external callback to sign any specific transaction is handled on the end-user(developer)'s side.

the frontend module now uses the local sdk tarball (for testing purposes), when we get to a stable version, we'd use the npm package directly.

* fix upload file function

* fix: failing deposit transactions

fixed an issue where deposit transactions always fail because of a wrong property pubKey, instead of pubkey in teh instruction data.

i've also added fn docs for the methods in the client.