fix: deposit-upload transaction flow
the previous approach we used where we compare the precomputed CID of a file to be uploaded was inefficient as the file is still uploaded to storacha regardless of a deposit being made.

now, this current approach ensures that a deposit (successful) is maded before the request to upload file(s) is made at all.