fix: deposit-upload transaction flow (#35)
* fix: deposit-upload transaction flow

the previous approach we used where we compare the precomputed CID of a file to be uploaded was inefficient as the file is still uploaded to storacha regardless of a deposit being made.

now, this current approach ensures that a deposit (successful) is made before the request to upload file(s) is made at all.

* fix: move deposit computation from the upload function

all the related deposit compute should be in the deposit function and the sdk should only call uploadFile when the transaction is confirmed.