fix: correct implementation for pre-computed CIDs of directories uploaded
the main problem was from the deposit function. i wasn't capturing the case of multiple files passed in the request, so the computeCid function allways falls back to calculating a hash for single file.

i've also refactored the computeDirectoryCID function such that the root CID is obtained properly via the directoryEncoderStream, quite similar to the way it was used in the upload-service.