feat: calculate backup expiration (#39)
when people upload data they want to store on Storacha for the duration they've specified, we should compute the expiration and include it in our db.

i noticed that the initial implementation did not include the date such an upload/deposit flow was initialized.

in this PR, i ensured that the data is included in the deposit table.

now that we have the expiration, we can decide how we want to enforce a rule for data that exceeds the specified duration(s).