fix: fixed an issue where upstash signature verification fails
for schedules created in the console, the body property is always optional, which was what i did. leaving it as is, hence an empty string was assigned to it when i triggered it manually.

initially i was decoding the request body with `JSON.stringify(req.body)` leading to the hash mismatch between the server and what upstash expects.

now, i assign an empty string if the body is empty.