feat: add extend_storage_duration instruction to contract
i've included a new extend_storage_duration instruction to handle renewals properly. turns out we were using create_deposit for renewals which tried to init a new account that already existed - solana kept yelling 'account already in use' at me.

in the contract, i added a new context type ExtendStorageDuration on existing deposits or uploads as we've decided to start calling them now. it mimics the same thing create_deposit does, but, specifically for extending teh storage duration.

i also included a constraint to prevent extensions from accounts/wallet addresses that did not initiate the upload initially — i think this may be an issue in the future, but let's find out first.

i was a bit confused after spending sometime in the backend module looking at how we read the program keypair and i discovered that i was using the same public key for deploying the program as the authorizing key, so i refactored it and also updated the respective READMEs to guide people. rookie move, mehnn!!