fix: fixed ui deployment issue related to vite
a lot of web3 packages are designed to run in the nodejs environment by default, and our sdk is not left out.

since vite, doesn't expose any of the server-side polyfills for node, we need to include the plugin to fix this error.

i've also updated the SDK readme for people who'd end up using it in a vite project.